Move deprecated get_geojson_observations and its helper functions back to node_api.py, with a DeprecationWarning
