+++ test/v1/test_node_api.py
@@ -5,6 +5,7 @@ from unittest.mock import patch
 
 import pyinaturalist
 from pyinaturalist.constants import API_V1_BASE_URL
+from pyinaturalist.node_api import get_geojson_observations
 from pyinaturalist.v1 import get_observation, get_user_by_id, get_users_autocomplete
 from test.conftest import load_sample_data
 
@@ -39,6 +40,37 @@ def test_get_users_autocomplete(requests_mock):
     assert first_result['created_at'] == datetime(2018, 4, 23, 17, 11, 14, tzinfo=tzutc())
 
 
+# TODO: Deprecate get_geojson_observations and move to pyinaturalist-convert
+def test_get_geojson_observations(requests_mock):
+    requests_mock.get(
+        f'{API_V1_BASE_URL}/observations',
+        json=load_sample_data('get_observation.json'),
+        status_code=200,
+    )
+
+    geojson = get_geojson_observations(id=16227955)
+    feature = geojson['features'][0]
+    assert feature['geometry']['coordinates'] == [4.360086, 50.646894]
+    assert feature['properties']['id'] == 16227955
+    assert feature['properties']['taxon_id'] == 493595
+
+
+def test_get_geojson_observations__custom_properties(requests_mock):
+    requests_mock.get(
+        f'{API_V1_BASE_URL}/observations',
+        json=load_sample_data('get_observation.json'),
+        status_code=200,
+    )
+
+    properties = ['taxon_name', 'taxon_rank']
+    geojson = get_geojson_observations(id=16227955, properties=properties)
+    feature = geojson['features'][0]
+    assert feature['properties']['taxon_name'] == 'Lixus bardanae'
+    assert feature['properties']['taxon_rank'] == 'species'
+    assert 'id' not in feature['properties'] and 'taxon_id' not in feature['properties']
+
+
+# TODO: Deprecate globals in favort of API client settings
 @patch.dict(os.environ, {'GITHUB_REF': 'refs/heads/main'}, clear=True)
 def test_user_agent(requests_mock):
     requests_mock.get(
